README: Refer to protocol in specific phases instead of directories
We now have a bunch of protocol phases which we can refer to. Mentioning
protocols by the implementation detail of in which directory they are is
a bit roundabout.

This change makes sure that protocol phases are used when possible and
mentions the directory structure exhaustively in a single place.

Signed-off-by: Sebastian Wick <sebastian.wick@redhat.com>