README: Suggested a new development model for experimental protocols
The current model requires backwards compatible changes to experimental
protocols to bump the protocol minor versions and incompatible changes
to bump the major version.

This is essentially the same exact same process as staging and stable,
the only difference being the expectation that major version bumps
should occur.

In the development process of the color-management protocol we
introduced the experimental protocol process, but it worked differently
from what is documented here.

Protocol changes were done without bumping major or minor versions,
making it painless to do any change and encourage major changes when
they are useful.

To make experimental implementations possible and interoperable even
when work on the protocol continues, snapshots of the protocol were made
with an increasing major version.

This process has a track record of actually working in practice.

We still have no experimental protocols currently, so let's take the
opportunity to adjust things now.

The actual changes to the README introduce the concept of a `devel`
protocol which is not meant to be consumed, but instead reflects the
latest developments.

Signed-off-by: Sebastian Wick <sebastian.wick@redhat.com>