Start an input-method protocol
This commit introduces an experimental input-method protocol as an exact copy of the fle describing the unofficial zwp_input_method_v2 from squeekboard.

It's also supported by wlroots and smithay.

This protocol is the counterpart to text-input-v3. It gives the compositor a standard way to outsource the handling of the input method.

Signed-off-by: Dorota Czaplejewicz