ext-image-copy-capture: Introduce preferred_done event
ext_image_copy_capture_session_v1 sends all supported buffer constraints
which the client combines with its own internal constraints to allocate
an appropriate buffer for use with capture frames. The buffer
constraints have no order.

The compositor should include a wide set of buffer constraints for
compatibility and to enable various use-cases. The client should work
with a wide set of buffer constraints to work in different environments.

When both compositor and client supports a wide set of buffer
constraints for compatibility and the client does not have a strong
preference of its own, a suboptimal or in some cases outright
nonsensical format may end up being selected simply because it was the
first result of an unordered intersection, which could negatively impact
performance or color resolution.

In this case, the client would benefit from including compositor
preference in the consideration. Compositor preference could for example
be the format of its current scanout or composition buffer.

Introduce proferred_done to delimit preferred buffer formats from
remaining compatible buffer formats.

Signed-off-by: Kenny Levinsen <kl@kl.wtf>