staging: xdg-decoration-v1
This protocol brings the unstable xdg-decoration-v1 into the modern age, and allows compositors such as Mutter, and toolkits like GTK, to support (at least a subset of) server-side decorations.

Compositors are not forced to advertise support for the full range of capabilities in this protocol, and for the capabilities they do expose, clients get full control over what gets drawn on their windows.

Co-authored-by: Neal Gompa <neal@gompa.dev>