single-pixel-buffer: remove note about interpretation
A wl_buffer is always a dumb collection of color channels whose
interpretation is determined ad-hoc at the point of use. For example, is
is possible for the same buffer to be attached two wl_surfaces
concurrently:

- The first surface interprets the color channels as sRGB with
  pre-multiplied alpha.
- The second surface interprets the color channels as BT.2020/PQ with
  straight alpha.

Therefore the interpretation is not a property of the wl_buffer itself.
Factory interfaces for wl_buffer object should therefore only talk about
how the electrical values are communicated between clients and
compositors.

Signed-off-by: Julian Orth <ju.orth@gmail.com>