wp-swapchain-lock-v1: new protocol
> This protocol allows application components to ensure that there is
> only one swapchain for a wl_surface.
>
> Components using this protocol should ensure that they never
>
> - make wl_surface.attach requests,
> - create a wp_linux_drm_syncobj_surface_v1 extension,
> - create a wp_commit_timer_v1 extension,
> - create a wp_fifo_v1 extension,
> - create a wp_color_management_surface_v1 extension, or
> - create a wp_color_representation_surface_v1 extension
>
> for a wl_surface unless they have acquired the lock. Furthermore,
> they should destroy all such extension objects before releasing the
> lock.

In particular, it allows vulkan drivers to implement the requirement
that vkCreateSwapchainKHR returns VK_ERROR_NATIVE_WINDOW_IN_USE_KHR if
another swapchain for the surface already exists. Currently no error is
returned and the application is disconnected because the driver tries to
create duplicate extension objects.

It might also help to fix persistent issues such as
https://github.com/obsproject/obs-studio/issues/11022

The word _swapchain_ in the protocol name is used to refer to the
abstract concept of the application component that is in charge of
presentation.

Signed-off-by: Julian Orth <ju.orth@gmail.com>